Support newer versions of hermes
The npm hermes-engine appears to be behind the actual hermes releases
(as of writing, stuck on v11 where v13 is available). Switch the hermes
downloader to look up the version via the github releases API, and fix
up the URL predictor to support a change in the download URL format.